src/app/shared/components/custom-select/custom-select.component.spec.ts
import { ComponentFixture, TestBed, async } from '@angular/core/testing';
import { CustomSelectComponent } from './custom-select.component';
import { By } from '@angular/platform-browser';
import { CUSTOM_ELEMENTS_SCHEMA, DebugElement, NO_ERRORS_SCHEMA } from '@angular/core';

describe('CustomSelectComponent', () => {
  let component: CustomSelectComponent;
  let fixture: ComponentFixture<CustomSelectComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [CustomSelectComponent],
      schemas: [CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA],
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(CustomSelectComponent);
    component = fixture.componentInstance;
    component.options = ['Option 1', 'Option 2', 'Option 3'];
    fixture.detectChanges();
  });

  it('should create the component', () => {
    expect(component).toBeTruthy();
  });

  it('should have initial state', () => {
    expect(component.showOptions).toBe(false);
    expect(component.selectedOption).toBe('Option 1');
  });
});